+++ src/file_entry.c
@@ -27,6 +27,7 @@
 #include "types.h"
 #include "utils.h"
 #include "options.h"
+#include "partition.h"
 #include "file_entry.h"
 
 /* stat(2) */
@@ -169,7 +170,7 @@ fpart_hook(const char *cmd, const struct program_options *options,
 
         if(options->verbose >= OPT_VERBOSE)
             fprintf(stderr, "Executing pre-part #%ju hook: '%s'\n",
-                display_index(*live_partition_index, options), cmd);
+                adapt_partition_index(*live_partition_index, options), cmd);
 
         /* FPART_HOOKTYPE (pre-part) */
         malloc_size = strlen(env_fpart_hooktype_name) + 1 +
@@ -192,7 +193,7 @@ fpart_hook(const char *cmd, const struct program_options *options,
 
         if(options->verbose >= OPT_VERBOSE)
             fprintf(stderr, "Executing post-part #%ju hook: '%s'\n",
-                display_index(*live_partition_index, options), cmd);
+                adapt_partition_index(*live_partition_index, options), cmd);
 
         /* FPART_HOOKTYPE (post-part) */
         malloc_size = strlen(env_fpart_hooktype_name) + 1 +
@@ -228,13 +229,13 @@ fpart_hook(const char *cmd, const struct program_options *options,
     /* FPART_PARTNUMBER */
     if(live_partition_index != NULL) {
         malloc_size = strlen(env_fpart_partnumber_name) + 1 +
-            get_num_digits(display_index(*live_partition_index, options)) + 1;
+            get_num_digits(adapt_partition_index(*live_partition_index, options)) + 1;
         if_not_malloc(env_fpart_partnumber_string, malloc_size,
             retval = 1;
             goto cleanup;
         )
         snprintf(env_fpart_partnumber_string, malloc_size, "%s=%ju",
-            env_fpart_partnumber_name, display_index(*live_partition_index, options));
+            env_fpart_partnumber_name, adapt_partition_index(*live_partition_index, options));
         if(push_env(env_fpart_partnumber_string, &envp) != 0) {
             retval = 1;
             goto cleanup;
@@ -370,7 +371,10 @@ cleanup:
     return (retval);
 }
 
-/* Print or add a file entry (redirector) */
+/* Print or add a file entry (redirector)
+   - returns (0) if entry has been added
+   - returns (1) if entry has been skipped (option -S)
+   - returns (-1) if error */
 int
 handle_file_entry(struct file_entry **head, char *path, fsize_t size,
     struct program_options *options)
@@ -383,7 +387,10 @@ handle_file_entry(struct file_entry **head, char *path, fsize_t size,
         return (add_file_entry(head, path, size, options));
 }
 
-/* Print a file entry */
+/* Print a file entry
+   - returns (0) if entry has been added
+   - returns (1) if entry has been skipped (option -S)
+   - returns (-1) if error */
 int
 live_print_file_entry(char *path, fsize_t size,
     struct program_options *options)
@@ -396,6 +403,19 @@ live_print_file_entry(char *path, fsize_t size,
     char *ln_term = (options->out_zero == OPT_OUT0) ? "\0" : "\n";
     int split = 0;
 
+    /* option -S: skip files bigger than maximum partition size (option -s)
+       and print them to stdout in hardcoded pseudo-partition 'S' ('S'kipped).
+       Preloading and overloading are used here too */
+    if(options->skip_big == OPT_SKIPBIG) {
+        fsize_t needed_part_size = options->preload_size +
+            round_num(size + options->overload_size, options->round_size);
+        if(needed_part_size > options->max_size) {
+            fprintf(stdout, "S (%ju): %s\n", size, path);
+            fflush(stdout);
+            return (1);
+        }
+    }
+
     /* beginning of a new partition */
     if(live_status.partition_num_files == 0) {
 start_part:
@@ -407,12 +427,12 @@ start_part:
         if(out_template != NULL) {
             /* compute live_status.filename "out_template.i\0" */
             size_t malloc_size = strlen(out_template) + 1 +
-                get_num_digits(display_index(live_status.partition_index, options)) + 1;
+                get_num_digits(adapt_partition_index(live_status.partition_index, options)) + 1;
             if_not_malloc(live_status.filename, malloc_size,
-                return (1);
+                return (-1);
             )
             snprintf(live_status.filename, malloc_size, "%s.%ju", out_template,
-                display_index(live_status.partition_index, options));
+                adapt_partition_index(live_status.partition_index, options));
         }
 
         /* execute pre-partition hook */
@@ -431,7 +451,7 @@ start_part:
                     strerror(errno));
                 free(live_status.filename);
                 live_status.filename = NULL;
-                return (1);
+                return (-1);
             }
         }
     }
@@ -450,7 +470,7 @@ start_part:
     if(out_template == NULL) {
         /* no template provided, just print to stdout */
         fprintf(stdout, "%ju (%ju): %s\n",
-            display_index(live_status.partition_index, options), size,
+            adapt_partition_index(live_status.partition_index, options), size,
             path);
     }
     else {
@@ -461,7 +481,7 @@ start_part:
             fprintf(stderr, "%s\n", strerror(errno));
             /* do not close(livefd) and free(live_status.filename) here because
                it will be useful and free'd in uninit_file_entries() below */
-            return (1);
+            return (-1);
         }
     }
 
@@ -477,8 +497,8 @@ start_part:
 end_part:
         /* display added partition */
         if(options->verbose >= OPT_VERBOSE)
-            fprintf(stderr, "Filled part #%ju: size = %jd, %ju file(s)\n",
-                display_index(live_status.partition_index, options),
+            fprintf(stderr, "Filled part #%ju: size = %ju, %ju file(s)\n",
+                adapt_partition_index(live_status.partition_index, options),
                 live_status.partition_size,
                 live_status.partition_num_files);
 
@@ -522,6 +542,8 @@ end_part:
 /* Add a file entry to a double-linked list of file_entries
    - if head is NULL, creates a new file entry ; if not, chains a new file
      entry to it
+   - returns (0) if entry has been added
+   - returns (-1) if error
    - returns with head set to the newly added element */
 int
 add_file_entry(struct file_entry **head, char *path, fsize_t size,
@@ -539,7 +561,7 @@ add_file_entry(struct file_entry **head, char *path, fsize_t size,
     previous = *current;
 
     if_not_malloc(*current, sizeof(struct file_entry),
-        return (1);
+        return (-1);
     )
 
     /* set head on first call */
@@ -552,7 +574,7 @@ add_file_entry(struct file_entry **head, char *path, fsize_t size,
     if_not_malloc((*current)->path, malloc_size,
         free(*current);
         *current = previous;
-        return (1);
+        return (-1);
     )
     snprintf((*current)->path, malloc_size, "%s", path);
     (*current)->size = size + options->overload_size;
@@ -792,10 +814,11 @@ add_directory:
                         curdir_size = -p->fts_errno;
                     }
                     /* add or display it */
-                    if(handle_file_entry
-                        (head, curdir_entry_path, curdir_size, options) == 0)
+                    int handled = handle_file_entry
+                        (head, curdir_entry_path, curdir_size, options);
+                    if(handled == 0)
                         (*count)++;
-                    else {
+                    else if(handled < 0) {
                         fprintf(stderr, "%s(): cannot add file entry\n",
                             __func__);
                         free(curdir_entry_path);
@@ -904,10 +927,11 @@ two-pass check to handle include and exclude options properly. */
                     continue;
 
                 /* add or display it */
-                if(handle_file_entry
-                    (head, p->fts_path, curfile_size, options) == 0)
+                int handled = handle_file_entry
+                    (head, p->fts_path, curfile_size, options);
+                if(handled == 0)
                     (*count)++;
-                else {
+                else if(handled < 0) {
                     fprintf(stderr, "%s(): cannot add file entry\n", __func__);
                     fts_close(ftsp);
                     return (1);
@@ -955,8 +979,8 @@ uninit_file_entries(struct file_entry *head, struct program_options *options)
         /* display added partition */
         if((options->verbose >= OPT_VERBOSE) &&
             (live_status.partition_num_files > 0))
-            fprintf(stderr, "Filled part #%ju: size = %jd, %ju file(s)\n",
-                display_index(live_status.partition_index, options),
+            fprintf(stderr, "Filled part #%ju: size = %ju, %ju file(s)\n",
+                adapt_partition_index(live_status.partition_index, options),
                 live_status.partition_size,
                 live_status.partition_num_files);
 
@@ -992,10 +1016,11 @@ uninit_file_entries(struct file_entry *head, struct program_options *options)
 /* Print a double-linked list of file_entries from head
    - if no filename template given, print to stdout */
 int
-print_file_entries(struct file_entry *head, pnum_t num_parts,
-    struct program_options *options)
+print_file_entries(struct file_entry *head, struct partition *part_head,
+    pnum_t num_parts, struct program_options *options)
 {
     assert(head != NULL);
+    assert(part_head != NULL);
     assert(num_parts > 0);
     assert(options != NULL);
 
@@ -1006,7 +1031,7 @@ print_file_entries(struct file_entry *head, pnum_t num_parts,
     if(out_template == NULL) {
         while(head != NULL) {
             fprintf(stdout, "%ju (%ju): %s\n",
-                display_index(head->partition_index, options),
+                adapt_partition_index(head->partition_index, options),
                 head->size, head->path);
             head = head->nextp;
         }
@@ -1015,48 +1040,67 @@ print_file_entries(struct file_entry *head, pnum_t num_parts,
 
     /* a template has been provided; to avoid opening too many files,
        open chunks of FDs and do as many passes as necessary */
-    struct file_entry *start = head;
-    pnum_t current_chunk = 0;           /* current chunk */
-    pnum_t current_file_entry = 0;      /* current file entry within chunk */
-
     assert(PRINT_FE_CHUNKS > 0);
 
-    while(((current_chunk * PRINT_FE_CHUNKS) + current_file_entry) <
-        num_parts) {
+    struct file_entry *start = head;
+    pnum_t current_chunk = 0;       /* current chunk */
+    pnum_t current_fd_index = 0;    /* current file index within chunk */
+
+    /* current -absolute- partition index */
+#define current_partition_index \
+    ((current_chunk * PRINT_FE_CHUNKS) + current_fd_index)
+
+    /* current_partition_index gets incremented by PRINT_FE_CHUNKS */
+    while(current_partition_index < num_parts) {
         int fd[PRINT_FE_CHUNKS]; /* our <files per chunk> file descriptors */
 
         /* open as necessary file descriptors as needed
            to print num_part partitions */
-        while((current_file_entry < PRINT_FE_CHUNKS) &&
-              (((current_chunk * PRINT_FE_CHUNKS) + current_file_entry) < num_parts)) {
+        while((current_fd_index < PRINT_FE_CHUNKS) &&
+              (current_partition_index < num_parts)) {
+            /* skip empty partition '0':
+               XXX No need to skip write() calls below, as no file entry should
+               be associated with that partition */
+            if((current_partition_index == 0) && (part_head->num_files == 0)) {
+#if defined(DEBUG)
+                fprintf(stderr, "%s(): skip creating empty partition '0'\n",
+                    __func__);
+#endif
+                fd[current_fd_index] = (-1); /* for close() calls below */
+                current_fd_index++;
+                continue;
+            }
+
             /* compute out_filename  "out_template.i\0" */
             char *out_filename = NULL;
             size_t malloc_size = strlen(out_template) + 1 + get_num_digits
-                ((current_chunk * PRINT_FE_CHUNKS) + current_file_entry) + 1;
+                (adapt_partition_index(current_partition_index, options)) + 1;
             if_not_malloc(out_filename, malloc_size,
                 /* close all open descriptors and return */
                 pnum_t i;
-                for(i = 0; i < current_file_entry; i++)
+                for(i = 0; i < current_fd_index; i++)
                      close(fd[i]);
                 return (1);
             )
             snprintf(out_filename, malloc_size, "%s.%ju", out_template,
-                (current_chunk * PRINT_FE_CHUNKS) + current_file_entry);
+                adapt_partition_index(current_partition_index, options));
 
-            if((fd[current_file_entry] =
+            if((fd[current_fd_index] =
                 open(out_filename, O_WRONLY|O_CREAT|O_TRUNC, 0660)) < 0) {
                 fprintf(stderr, "%s: %s\n", out_filename, strerror(errno));
                 free(out_filename);
                 /* close all open descriptors and return */
                 pnum_t i;
-                for(i = 0; i < current_file_entry; i++)
+                for(i = 0; i < current_fd_index; i++)
                      close(fd[i]);
                 return (1);
             }
             free(out_filename);
-            current_file_entry++;
+
+            current_fd_index++;
         }
 
+        /* write data to opened file descriptors */
         while(head != NULL) {
             if((head->partition_index >= (current_chunk * PRINT_FE_CHUNKS)) &&
                (head->partition_index < ((current_chunk + 1) * PRINT_FE_CHUNKS))) {
@@ -1081,7 +1125,7 @@ print_file_entries(struct file_entry *head, pnum_t num_parts,
         for(i = 0; (i < PRINT_FE_CHUNKS) && (((current_chunk * PRINT_FE_CHUNKS) + i) < num_parts); i++)
             close(fd[i]);
 
-        current_file_entry = 0;
+        current_fd_index = 0;
         current_chunk++;
     }