+++ src/fpart.c
@@ -142,6 +142,10 @@ usage(void)
     fprintf(stderr, "Live mode:\n");
     fprintf(stderr, "  -L\tlive mode: generate partitions during filesystem "
         "crawling\n");
+    fprintf(stderr, "  -S\tdo not pack files bigger than specified maximum "
+        "partition size\n");
+    fprintf(stderr, "    \tbut print them to stdout instead (needs -L, -s and no "
+        "-o)\n");
     fprintf(stderr, "  -w\tpre-partition hook: execute <cmd> at partition "
         "start\n");
     fprintf(stderr, "  -W\tpost-partition hook: execute <cmd> at partition "
@@ -183,9 +187,11 @@ handle_argument(char *argument, fnum_t *totalfiles, struct file_entry **head,
         )
 
         if(sscanf(argument, "%ju %[^\n]", &input_size, input_path) == 2) {
-            if(handle_file_entry(head, input_path, input_size, options) == 0)
+            int handled = handle_file_entry
+                (head, input_path, input_size, options);
+            if(handled == 0)
                 (*totalfiles)++;
-            else {
+            else if(handled < 0) {
                 fprintf(stderr, "%s(): cannot add file entry\n", __func__);
                 free(input_path);
                 return (1);
@@ -261,9 +267,9 @@ handle_options(struct program_options *options, int *argcp, char ***argvp)
     int ch;
     while((ch = getopt(*argcp, *argvp,
 #if defined(_HAS_FNM_CASEFOLD)
-        "hVn:f:s:i:ao:0evlby:Y:x:X:zZd:DELw:W:p:q:r:"
+        "hVn:f:s:i:ao:0evlby:Y:x:X:zZd:DELSw:W:p:q:r:"
 #else
-        "hVn:f:s:i:ao:0evlby:x:zZd:DELw:W:p:q:r:"
+        "hVn:f:s:i:ao:0evlby:x:zZd:DELSw:W:p:q:r:"
 #endif
         )) != -1) {
         switch(ch) {
@@ -422,6 +428,9 @@ handle_options(struct program_options *options, int *argcp, char ***argvp)
             case 'L':
                 options->live_mode = OPT_LIVEMODE;
                 break;
+            case 'S':
+                options->skip_big = OPT_SKIPBIG;
+                break;
             case 'w':
             {
                 /* check for empty argument */
@@ -571,6 +580,15 @@ handle_options(struct program_options *options, int *argcp, char ***argvp)
         }
         if (options->max_size == 0)
             options->max_size = INTMAX_MAX;
+
+    /* option -S (needs '-L', '-s' and no '-o') */
+    if((options->skip_big == OPT_SKIPBIG) &&
+        ((options->live_mode == OPT_NOLIVEMODE) ||
+        (options->out_filename != NULL) ||
+        (options->max_size == DFLT_OPT_MAX_SIZE))) {
+        fprintf(stderr,
+            "Option -S can only be used with options -L and -s (without -o).\n");
+        return (FPART_OPTS_USAGE | FPART_OPTS_NOK | FPART_OPTS_EXIT);
     }
 
     if((options->in_filename == NULL) && (*argcp <= 0)) {
@@ -648,7 +666,7 @@ int main(int argc, char **argv)
         bzero(line, MAX_LINE_LENGTH);
         while(fgets(line, MAX_LINE_LENGTH, in_fp) != NULL) {
             /* replace '\n' with '\0' */
-            if ((line_end_p = strchr(line, '\n')) != NULL)
+            if((line_end_p = strchr(line, '\n')) != NULL)
                 *line_end_p = '\0';
 
             if(handle_argument(line, &totalfiles, &head, &options) != 0) {
@@ -810,7 +828,7 @@ int main(int argc, char **argv)
         fprintf(stderr, "Writing output lists...\n");
 
     /* print file entries */
-    print_file_entries(head, num_parts, &options);
+    print_file_entries(head, part_head, num_parts, &options);
 
     if(options.verbose >= OPT_VERBOSE)
         fprintf(stderr, "Cleaning up...\n");