+++ src/partition.h
@@ -44,9 +44,11 @@ struct partition {
 
 int add_partitions(struct partition **head, pnum_t num_parts,
     struct program_options *options);
+int remove_partition(struct partition *part);
 void uninit_partitions(struct partition *head);
 pnum_t find_smallest_partition_index(struct partition *head);
 struct partition * get_partition_at(struct partition *head, pnum_t index);
+pnum_t adapt_partition_index(pnum_t index, const struct program_options *options);
 void print_partitions(struct partition *head, struct program_options *options);
 
 #endif /* _PARTITION_H */