+++ src/utils.c
@@ -380,7 +380,7 @@ str_to_uintmax(const char *str, const unsigned char handle_multiplier)
 
         uintmax_t orig_val = val;
         /* more than one character remain or invalid multiplier specified */
-        if ((*(endptr + 1) != '\0') ||
+        if((*(endptr + 1) != '\0') ||
             (multiplier = char_to_multiplier(*endptr)) == 0) {
             fprintf(stderr, "%s(): %s\n", __func__, "unknown unit provided");
             return (0);
@@ -560,24 +560,4 @@ push_env(char *str, char ***env)
     *env = new_env;
 
     return (0);
-}
-
-/* Adapt partition index for display, regarding program options
-   - returns an index suitable for user display */
-pnum_t
-display_index(pnum_t index, const struct program_options *options)
-{
-    assert(options != NULL);
-
-    pnum_t display_offset = 1;
-
-    /* compute displayed index offset: fpart always displays partitions starting
-       from '1' (but they internally start from '0'). Partition '0' -used for
-       large files that do not fit in regular partitions- only appears when
-       option '-s' is passed *and* non-live mode is used */
-    if((options->max_size != DFLT_OPT_MAX_SIZE) &&
-       (options->live_mode == OPT_NOLIVEMODE))
-        display_offset = 0;
-
-    return (index + display_offset);
 }